requirements.txt
gradio>=4.20.0
torch>=2.1.0
torchvision>=0.16.0
monai>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
matplotlib>=3.7.0
scipy>=1.11.0
